requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
asyncio==3.4.3
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
et-xmlfile==1.1.0
exrex==0.11.0
filelock==3.13.1
frozenlist==1.4.1
googlesearch-python==1.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inquirerpy==0.3.4
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.0.0
multidict==6.0.4
numpy==1.26.3
openpyxl==3.0.10
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parsel==1.8.1
pfzy==0.3.4
priority==2.0.0
prompt-toolkit==3.0.1
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pyfiglet==1.0.2
Pygments==2.17.2
PyInquirer==1.0.3
PyJWT==2.8.0
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nmap==0.7.1
pytz==2023.3.post1
queuelib==1.6.2
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
requests-futures==1.0.1
Scrapy==2.11.0
selenium==4.16.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
stem==1.8.2
tldextract==5.1.1
torrequest==0.1.0
trio==0.23.2
trio-websocket==0.11.1
twilio==8.11.0
Twisted==22.10.0
twisted-iocpsupport==1.0.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
w3lib==2.1.2
wcwidth==0.2.12
wsproto==1.2.0
yarl==1.9.4
zope.interface==6.1
